refactor: itemList 훅 정리
