+++ src/hooks/itemList/useItemList.ts
@@ -1,30 +1,11 @@
-// src/hooks/useItemList.ts
-import { useState } from "react";
-import { useMemo } from "../../hooks";
-import { useThemeState } from "../../context";
-import { Item } from "../../types";
+import { useItemData } from "./useItemData";
+import { useItemFilter } from "./useItemFilter";
+import { useItemStats } from "./useItemStats";
 
-export function useItemList(items: Item[]) {
-  const [filter, setFilter] = useState("");
-  const theme = useThemeState();
-
-  const filteredItems = useMemo(() => {
-    if (filter === "") return items;
-    const lower = filter.toLowerCase();
-    return items.filter(
-      (item) =>
-        item.name.toLowerCase().includes(lower) ||
-        item.category.toLowerCase().includes(lower)
-    );
-  }, [items, filter]);
-
-  const totalPrice = useMemo(() => {
-    return filteredItems.reduce((sum, item) => sum + item.price, 0);
-  }, [filteredItems]);
-
-  const averagePrice = useMemo(() => {
-    return Math.round(totalPrice / filteredItems.length) || 0;
-  }, [totalPrice, filteredItems.length]);
+export function useItemList() {
+  const { items, addItems } = useItemData();
+  const { filter, setFilter, filteredItems } = useItemFilter(items);
+  const { totalPrice, averagePrice } = useItemStats(filteredItems);
 
   return {
     filter,
@@ -32,6 +13,6 @@ export function useItemList(items: Item[]) {
     filteredItems,
     totalPrice,
     averagePrice,
-    theme,
+    addItems,
   };
 }
