+++ src/pages/HomePage.tsx
@@ -1,21 +1,9 @@
-import { useState } from "react";
-import { useCallback } from "../hooks";
 import { useThemeState } from "../context";
 import { ItemList, ComplexForm, NotificationSystem } from "../components";
-import { generateItems } from "../utils";
 import { Header } from "../components";
 
 const HomePage: React.FC = () => {
   const theme = useThemeState();
-
-  const [items, setItems] = useState(() => generateItems(1000));
-
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
-  }, []);
 
   return (
     <div
@@ -25,7 +13,7 @@ const HomePage: React.FC = () => {
       <div className="container mx-auto px-4 py-8">
         <div className="flex flex-col md:flex-row">
           <div className="w-full md:w-1/2 md:pr-4">
-            <ItemList items={items} onAddItemsClick={addItems} />
+            <ItemList />
           </div>
           <div className="w-full md:w-1/2 md:pl-4">
             <ComplexForm />
